"Update email configurations and add getAdminEmails function for reusability"
This commit updates the email configurations in several files to use environment variables for the email user and password. It also adds a new function, getAdminEmails, in the adminEmails.js file for retrieving a list of admin email addresses. This function is then used in other files to send emails to multiple admin recipients. Additionally, the commit includes some minor reformatting and code cleanup.